Ready to accomodate new screen Can navigate from screen 1 to screen 2
