+++ eartest/lib/screen2.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+import 'package:audioplayers/audioplayers.dart';
+
+class screen2 extends StatelessWidget{
+  @override
+  Widget build(BuildContext context) {
+     return MaterialApp(
+       home: Scaffold(
+
+         //Implement screen 2 here
+
+       ),
+    );
+  }
+
+}
